Performance hacks: index pagination, persistent _internal
This are some quick hacks that make datasette more performant
in the face of lots and giant DBs. This adds pagination to the
index view (add ?page=0, ?page=1, etc to the index page to
get next pages). This also cuts down the number of tables that
will be iterated over per DB on the index page.

Secondarily, this makes the _internal database persistent at
./secrets/_internal.db. This *drastically* improves initial page
loading after a reboot, as creating the internal DB can take quite
some time with lots of dbs/large dbs.